#352 Merge annotators scripts and move to tools
