+++ avro/src/test/java/dev/hardwood/avro/AvroRowReaderTest.java
@@ -177,7 +177,7 @@ class AvroRowReaderTest {
 
     @Test
     void readShreddedVariantColumn() throws Exception {
-        // variant_shredded_test.parquet (generated by simple-datagen.py) has a
+        // variant_shredded_test.parquet (generated by tools/simple-datagen.py) has a
         // VARIANT-annotated group with {metadata, value, typed_value:int64} and
         // four rows exercising the distinct reassembly outcomes. The Avro view
         // follows parquet-java's AvroParquetReader shape: a two-field
