+++ tools/simple-datagen.py
@@ -12,11 +12,7 @@ from datetime import datetime, date, time, timezone
 from decimal import Decimal
 import uuid
 
-import struct
-
-from parquet_bson_annotation import annotate_column_as_bson
-from parquet_interval_annotation import annotate_column_as_interval
-from parquet_variant_annotation import annotate_group_as_variant
+from parquet_annotators import annotate_column_as_bson, annotate_group_as_variant
 
 # Plain encoding with no compression (for Milestone 1)
 # Create a simple table with NO nulls first, explicitly marking fields as non-nullable
